requirements.txt
pytorch
torchvision
numpy
pillow
torchvision